Move rfxtrx base entity to separate module
