+++ homeassistant/components/rfxtrx/siren.py
@@ -14,13 +14,9 @@ from homeassistant.helpers.entity import Entity
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 from homeassistant.helpers.event import async_call_later
 
-from . import (
-    DEFAULT_OFF_DELAY,
-    DeviceTuple,
-    RfxtrxCommandEntity,
-    async_setup_platform_entry,
-)
+from . import DEFAULT_OFF_DELAY, DeviceTuple, async_setup_platform_entry
 from .const import CONF_OFF_DELAY
+from .entity import RfxtrxCommandEntity
 
 SECURITY_PANIC_ON = "Panic"
 SECURITY_PANIC_OFF = "End Panic"
